exception handling with custom responses

+++ src/main/java/com/wego/cstms/service/ContentService.java
@@ -2,6 +2,8 @@ package com.wego.cstms.service;
 
 
 import com.wego.cstms.dto.mapper.CourseContentMapper;
+import com.wego.cstms.exceptions.EntityType;
+import com.wego.cstms.exceptions.MSException;
 import com.wego.cstms.persistence.Entities.Course;
 import com.wego.cstms.persistence.Entities.CourseContent;
 import com.wego.cstms.persistence.repositories.CourseContentRepository;
@@ -12,17 +14,20 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ContentService {
 
     private final CourseContentRepository courseContentRepository;
     private final CourseRepository courseRepository;
+    private final MSException msException;
 
     @Autowired
-    public ContentService(CourseContentRepository courseContentRepository, CourseRepository courseRepository) {
+    public ContentService(CourseContentRepository courseContentRepository, CourseRepository courseRepository, MSException msException) {
         this.courseContentRepository = courseContentRepository;
         this.courseRepository = courseRepository;
+        this.msException = msException;
     }
 
     public List<CourseContentDto> getCoursesAllContents(int id) {
@@ -33,17 +38,24 @@ public class ContentService {
         return courseContents;
     }
 
-    public String getDownloadPath(int courseId, int contentId){
+    public String[] getDownloadPath(int courseId, int contentId){
         CourseContent courseContent = courseContentRepository.findById(contentId).get();
-        return courseId+"/"+courseContent.getFilename();
+        String filename = courseContent.getFilename();
+        String filepath = courseId+"/"+filename;
+        return new String[]{filename, filepath};
     }
 
 
-    public void addCourseContent(CourseContentDto courseContentDto, int courseId) {
+    public CourseContentDto addCourseContent(CourseContentDto courseContentDto, int courseId) {
+        Optional<Course> course = courseRepository.findById(courseId);
+        if (course.isPresent()){
+            CourseContent courseContent = CourseContentMapper.toCourseContent(courseContentDto);
+            courseContent.setCourse(course.get());
+            courseContentRepository.save(courseContent);
+            return courseContentDto;
+        }
+        throw msException.EntityNotFoundException(EntityType.COURSE,courseId);
 
-        CourseContent courseContent = CourseContentMapper.toCourseContent(courseContentDto);
-        courseContent.setCourse(courseRepository.findById(courseId).get());
-        courseContentRepository.save(courseContent);
     }
 //
 //    public void updateCourseContent(ContentDto contentDto) {
@@ -51,8 +63,17 @@ public class ContentService {
 //        contentRepository.save(new Content(contentDto));
 //    }
 //
-    public void deleteCourseContent(Integer courseContentId) {
-        courseContentRepository.deleteById(courseContentId);
+    public String deleteCourseContent(Integer courseId,Integer courseContentId) {
+        if (courseRepository.existsById(courseId)){
+            Optional<CourseContent> courseContent = courseContentRepository.findById(courseId);
+            if(courseContent.isPresent()){
+                courseContentRepository.deleteById(courseContentId);
+                return String.format("Course Content:%s Deleted Successfully",courseContent.get().getFilename());
+            }
+            throw msException.EntityNotFoundException(EntityType.COURSE_CONTENT,courseId);
+        }
+        throw msException.EntityNotFoundException(EntityType.COURSE,courseId);
+
     }
 
     public void deleteAllCourseContent(Integer courseContentId) {
